#44 create seat file and import it at room
